+++ src/shared/components/Badge/Badge.module.scss
@@ -7,17 +7,14 @@
   align-items: center;
   border-radius: 2rem;
   gap: 0.4rem;
-  //폰트
-  @include SpoqaHanSansNeo-Regular {
-    font-size: 1.4rem;
-    font-weight: 700;
-  }
+  @include sopoqa-14-700;
 }
-.mobile {
+.responsive {
   @include mobile {
     @include sopoqa-12-400;
   }
 }
+
 .blue {
   background: $color-blue-10;
   color: $color-blue-20;
